requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
pytesseract
pdf2image
numpy
Pillow
PyMuPDF